Introduce a new StepMode
Why:
Increment by value -> no good when zooming
Increment by pixel -> no good when zooming
subdivide -> depending on data chooses ticks at wired locations e.g (1.3 , 2.3, 3,3 instead of 1, 2, 3)

Workaround:
When you know your data: supply an array of predefined increments (by value) for StepModel to choose from
to best fit the desired number of lines. For example:
Start zoomed out with ticks every 100 and as you zoom in switch to 50,10,1
