+++ androidplot-core/src/main/java/com/androidplot/xy/StepModel.java
@@ -24,26 +24,10 @@ public class StepModel {
     public StepModel(StepMode mode, double value) {
         setMode(mode);
         setValue(value);
-        setSteps(null);
-    }
-
-    public StepModel(double[] increments, double numLines) {
-        setMode(StepMode.INCREMENT_BY_FIT);
-        setValue(numLines);
-        setSteps(increments);
     }
 
     private StepMode mode;
-    private double value;   // increment by x value, pixels or number of sub division
-    private double[] steps; // for fit mode: possible increments (by value) to choose from
-
-    public double[] getSteps() {
-        return steps;
-    }
-
-    public void setSteps(double[] steps) {
-        this.steps = steps;
-    }
+    private double value;
 
     public StepMode getMode() {
         return mode;
@@ -59,34 +43,5 @@ public class StepModel {
 
     public void setValue(double value) {
         this.value = value;
-    }
-
-    /*
-    depending on the currently displayed range (by value) choose increment so that
-    the number of lines is closest to value
-     */
-    public double getFitValue(double range) {
-
-        // no possible increments where supplied (e.g. switched into this mode without calling setSteps(...)
-        // TODO: throw exception this should not be done
-        if (steps == null)
-            return getValue();
-
-        double curStep = steps[0];
-
-        double oldDistance = Math.abs((range / curStep)-value );
-
-        // determine which step size comes closest to the desired number of steps
-        for (double step : steps) {
-
-            double newDistance = Math.abs((range / step)-value );
-
-            // closer than previos stepping?
-            if (newDistance < oldDistance){
-                curStep = step;
-                oldDistance = newDistance;
-            }
-        }
-        return curStep;
     }
 }
